Merge in EPACDX/cdx-oaqps-cef from development to master

* commit '6b201fb2b4d70eed9b01581f47317cb84a9edc93':
  CEF-1363 adding checking for null facility source type code
  CEF-1367 adding QA check to prevent PS TS processes or units from having status year prior to previously submitted OP status year for non landfill facilities
  [CEF-1369] Updating Cypress script to work with more browsers
  adding facility enforcer to json export
  CEF-1266 update to exclude trailing decimal zeros in pattern validation
  CEF-1266 update to fix calculating avg percent control to exclude null
  update to retrieve mfr by program sys code and agency id
  CEF-1359 update retrieve alt facility id from facility site
  CEF-1359 update query to retrieve mfr info by eis program id and alt facility id
  update emissions unit validator
  CEF-1266 change double to bigdecimal
  CEF-1359 overwrite BU facility info with MFR info
  [CEF-1362] Adding missing EF variables on emission screen

+++ cef-web/src/main/java/gov/epa/cef/web/domain/Control.java
@@ -1,5 +1,6 @@
 package gov.epa.cef.web.domain;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 import java.util.ArrayList;
 import java.util.List;
@@ -42,10 +43,10 @@ public class Control extends BaseAuditEntity {
     private String upgradeDescription;
 
     @Column(name = "percent_capture", precision = 4, scale = 1)
-    private Double percentCapture;
+    private BigDecimal percentCapture;
 
     @Column(name = "percent_control", precision = 6, scale = 3)
-    private Double percentControl;
+    private BigDecimal percentControl;
     
     @Column(name = "number_operating_months", precision = 2, scale = 0)
     private Short numberOperatingMonths;
@@ -134,19 +135,19 @@ public class Control extends BaseAuditEntity {
         this.description = description;
     }
 
-    public Double getPercentCapture() {
+    public BigDecimal getPercentCapture() {
         return percentCapture;
     }
 
-    public void setPercentCapture(Double percentCapture) {
+    public void setPercentCapture(BigDecimal percentCapture) {
         this.percentCapture = percentCapture;
     }
 
-    public Double getPercentControl() {
+    public BigDecimal getPercentControl() {
         return percentControl;
     }
 
-    public void setPercentControl(Double percentControl) {
+    public void setPercentControl(BigDecimal percentControl) {
         this.percentControl = percentControl;
     }
